Start of project - 1st Commit
This is the start of the POE - 11/04/2024.

I already created and finished the part 1 of this project, and even some of part 2. I had a GitHub repository for it and everything, made all the commits, and had a console app providing the necessary functionality.

I decided to start from scratch again, because I firstly created a Console App with a .NET 8, and we are required to make a .NET 4.8, and secondly, I think I might be able to make it more professional.