Added a Recipe class
I created a class called Recipe, which will consist of its attributes relating to the recipe requirements such as, the ingredients, the steps, the name of the recipe, and the factor that the recipe will be scaled by.

I added some functions with no functionality to help brainstorm, the flow of logic for my code.