Added functionality to DisplayRecipe() and CreateNewRecipe()
In the Program class, I created a test sample, to ensure that my DisplayRecipe() function works. With the use of a pre-set sample, it just makes for efficient work as I don't need to manually enter a recipe in to see if I am getting the desired output.

I also added functionality to CreateNewRecipe(), which takes in user input, and stores it in the memory of the object.