Added Ingredients class and edited code to suit new changes
I created an Ingredients class, with the properties: ingredient name, quantity, and the unit of measurement.

Then in my Recipe class, I changed the attribute that was previously holding the information regarding the ingredients, from a 2D string array, to a 1D array of Ingredients. Therefore I don't need to parse the double for quantity, every time I want to access it. I also think this method will help simplify the scale method.

I believe this method of having an ingredients class, is much more efficient and user-friendly to understand the code. Better standard of coding in my opinion.