Updated Flow of Logic for Recipe Manager
Some methods were being called unnecessarily, simplified the code so, all the necessary functions do not need to be called individually, rather when a major function such as "AddNewRecipe" is called, all the required functions are nested inside of the "AddNewRecipe" function.

This removed unnecessary lines of code that were redundant or not being used.