Optimized Application
I converted all of the Views, which are Windows, to Pages. This reduces the amount of resources required to run the application. Instead of windows continuously opening and closing, rather is it a frame in the Main Window, and the content of the frame is being updated by the menu buttons.